refractored out ratioTable to display ratio tables.
